backend/ingestion/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic-settings==2.2.1
pydantic==2.7.0

# Database
sqlalchemy==2.0.29
asyncpg==0.29.0
alembic==1.13.1

# File parsers
pymupdf==1.24.1              # PDF
pdfplumber==0.11.0           # PDF tables
python-docx==1.1.0           # Word
python-pptx==0.6.23          # PowerPoint
pandas==2.2.1                # CSV / Excel
openpyxl==3.1.2              # Excel
easyocr==1.7.1               # OCR images
Pillow==10.3.0               # Images
lxml==5.2.1                  # XML
python-magic==0.4.27         # File type detection

# NLP / AI extraction
spacy==3.7.4
sentence-transformers==2.7.0
keybert==0.8.4
httpx==0.27.0                # HTTP client inter-services

# Storage
boto3==1.34.84               # MinIO / S3

# Queue / cache
redis==5.0.4
celery==5.3.6

# Monitoring
prometheus-fastapi-instrumentator==6.1.0
